extract websocket methods to separate file
